Модуляция. Правки
